Split span-related code into its own package ('style')
