+++ indicator/src/main/java/com/fuzz/indicator/style/MigratoryRange.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator.text;
+package com.fuzz.indicator.style;
 
 import android.support.annotation.NonNull;
