+++ indicator/src/main/java/com/fuzz/indicator/style/MigratorySpan.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator.text;
+package com.fuzz.indicator.style;
 
 import android.support.annotation.NonNull;
+
+import com.fuzz.indicator.text.LayeredTextViewHolder;
 
 /**
  * Interface to be implemented by spans used in {@link LayeredTextViewHolder}.
